Cleaned the multiboot2 module.
There's still a point on the `ELFSymbols::at()` function which instead of directly using core::mem::size_of<Self>(), adds the supposed size of the struct.

This is because there is padding at the end : it could be resolved by using
a "packed" repr, but it makes many things unsafe so I prefer keeping it that way until I fiind a better solution.